+++ snakemake/dag.py
@@ -154,7 +154,7 @@ class DAG:
             except KeyError:
                 pass
 
-    def create_conda_envs(self, dryrun=False, forceall=False):
+    def create_conda_envs(self, dryrun=False, forceall=False, init_only=False):
         conda.check_conda()
         # First deduplicate based on job.conda_env_file
         jobs = self.jobs if forceall else self.needrun_jobs
@@ -168,7 +168,8 @@ class DAG:
             hash = env.hash
             self.conda_envs[env_file] = env
             if hash not in hash_set:
-                env.create(dryrun)
+                if not init_only:
+                    env.create(dryrun)
                 hash_set.add(hash)
 
     def pull_singularity_imgs(self, dryrun=False, forceall=False):
@@ -212,7 +213,7 @@ class DAG:
         jobids = self.workflow.persistence.external_jobids(job)
         if len(jobids) == 1:
             return jobids[0]
-        else:
+        elif len(jobids) > 1:
             raise WorkflowError(
                 "Multiple different external jobids registered "
                 "for output files of incomplete job {} ({}). This job "
@@ -385,7 +386,7 @@ class DAG:
             if os.path.realpath(shadow_output) == os.path.realpath(
                     real_output):
                 continue
-            logger.info("Moving shadow output {} to destination {}".format(
+            logger.debug("Moving shadow output {} to destination {}".format(
                 shadow_output, real_output))
             shutil.move(shadow_output, real_output)
         shutil.rmtree(job.shadow_dir)