+++ snakemake/remote/HTTP.py
@@ -6,6 +6,7 @@ __license__ = "MIT"
 import os
 import re
 import collections
+import shutil
 import email.utils
 from contextlib import contextmanager
 
@@ -152,14 +153,22 @@ class RemoteObject(DomainObject):
     def download(self, make_dest_dirs=True):
         with self.httpr(stream=True) as httpr:
             if self.exists():
+                # Find out if the source file is gzip compressed in order to keep
+                # compression intact after the download.
+                # Per default requests decompresses .gz files.
+                # More detials can be found here: https://stackoverflow.com/questions/25749345/how-to-download-gz-files-with-requests-in-python-without-decoding-it?noredirect=1&lq=1
+                # Since data transferred with HTTP compression need to be decompressed automatically
+                # check the header and decode if the content is encoded.
+                if not self.name.endswith(".gz") and httpr.headers["Content-Encoding"] == "gzip":
+                    # Decode non-gzipped sourcefiles automatically.
+                    # This is needed to decompress uncompressed files that are compressed
+                    # for the transfer by HTTP compression.
+                    httpr.raw.decode_content = True
                 # if the destination path does not exist
                 if make_dest_dirs:
                     os.makedirs(os.path.dirname(self.local_path), exist_ok=True)
-
                     with open(self.local_path, 'wb') as f:
-                        for chunk in httpr.iter_content(chunk_size=1024):
-                            if chunk: # filter out keep-alives
-                                f.write(chunk)
+                        shutil.copyfileobj(httpr.raw, f)
                     os.sync() # ensure flush to disk
             else:
                 raise HTTPFileException("The file does not seem to exist remotely: %s" % self.remote_file())