+++ CHANGELOG.md
@@ -1,5 +1,28 @@
 # Change Log
 
+# [4.5.0] - 2018-01-18
+### Added
+- iRODS remote provider
+### Changed
+- Bug fix in shell usage of scripts and wrappers.
+- Bug fixes for cluster execution, --immediate-submit and subworkflows.
+
+## [4.4.0] - 2017-12-21
+### Added
+- A new shadow mode (minimal) that only symlinks input files has been added.
+### Changed
+- The default shell is now bash on linux and maxOS. If bash is not installed, we fall back to sh. Previously, Snakemake used the default shell of the user, which defeats the purpose of portability. If the developer decides so, the shell can be always overwritten using shell.executable().
+- Snakemake now requires Singularity 2.4.1 at least (only when running with --use-singularity).
+- HTTP remote provider no longer automatically unpacks gzipped files.
+- Fixed various smaller bugs.
+
+## [4.3.1] - 2017-11-16
+### Added
+- List all conda environments with their location on disk via --list-conda-envs.
+### Changed
+- Do not clean up shadow on dry-run.
+- Allow R wrappers.
+
 ## [4.3.0] - 2017-10-27
 ### Added
 - GridFTP remote provider. This is a specialization of the GFAL remote provider that uses globus-url-copy to download or upload files.